Add schemas and routes for CRUD operations for submenus

+++ src/routes/menu.py
@@ -8,7 +8,8 @@ from loguru import logger
 
 from src.database import get_async_session
 from src.routes.abc_route import APIMenuRouter
-from src.schemas.menu import MenuOutSchema, MenuInSchema
+from src.schemas.base import BaseInSchema
+from src.schemas.menu import MenuOutSchema
 from src.schemas.response import ResponseSchema, ResponseForDeleteSchema
 from src.services.menu import MenuService
 from src.utils.exceptions import CustomApiException
@@ -41,7 +42,7 @@ async def get_menu_list(
     status_code=201,
 )
 async def create_menu(
-        new_menu: MenuInSchema,
+        new_menu: BaseInSchema,
         session: AsyncSession = Depends(get_async_session),
 ):
     """
@@ -49,6 +50,7 @@ async def create_menu(
     """
     new_menu = await MenuService.create(new_menu=new_menu, session=session)
     return new_menu
+
 
 @router.get(
     "/{menu_id}",
@@ -85,20 +87,21 @@ async def get_menu(
 )
 async def update_menu(
         menu_id: UUID,
-        data: MenuInSchema,
+        data: BaseInSchema,
         session: AsyncSession = Depends(get_async_session),
 ):
     """
     Обновление меню по id
     """
-    menu = await MenuService.update(menu_id=menu_id, data=data, session=session)
+    update_menu = await MenuService.get(menu_id=menu_id, session=session)
 
-    if menu:
-        return menu
-
-    raise CustomApiException(
-        status_code=HTTPStatus.NOT_FOUND, detail="menu not found"
-    )
+    if update_menu:
+        await MenuService.update(menu_id=menu_id, data=data, session=session)
+        return update_menu
+    else:
+        raise CustomApiException(
+            status_code=HTTPStatus.NOT_FOUND, detail="menu not found"
+        )
 
 
 @router.delete(
@@ -116,11 +119,12 @@ async def delete_menu(
     """
     Удаление меню по id
     """
-    res = await MenuService.delete(menu_id=menu_id, session=session)
+    delete_menu = await MenuService.get(menu_id=menu_id, session=session)
 
-    if res:
+    if delete_menu:
+        await MenuService.delete(menu_id=menu_id, session=session)
         return ResponseForDeleteSchema(message="The menu has been deleted")
-
-    raise CustomApiException(
-        status_code=HTTPStatus.NOT_FOUND, detail="menu not found"
-    )
+    else:
+        raise CustomApiException(
+            status_code=HTTPStatus.NOT_FOUND, detail="menu not found"
+        )
